allsoc: McSPI: fix MISRA C
- Fix MISRA C for mcspi examples and drivers

Fixes: MCUSDK-11070

+++ examples/drivers/mcspi/mcspi_loopback/mcspi_loopback.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2021 Texas Instruments Incorporated
+ *  Copyright (C) 2021-24 Texas Instruments Incorporated
  *
  *  Redistribution and use in source and binary forms, with or without
  *  modification, are permitted provided that the following conditions
@@ -47,6 +47,7 @@
 #include "ti_drivers_config.h"
 #include "ti_drivers_open_close.h"
 #include "ti_board_open_close.h"
+#include <drivers/hw_include/tistdtypes.h>
 
 #define APP_MCSPI_MSGSIZE                   (128U)
 #define APP_MCSPI_TRANSFER_LOOPCOUNT        (10U)
@@ -60,7 +61,7 @@ void *mcspi_loopback_main(void *args)
     uint32_t            i, j;
     int32_t             transferOK;
     MCSPI_Transaction   spiTransaction;
-    uint64_t            startTimeInUSec, elapsedTimeInUsecs;
+    uint32_t            startTimeInUSec, elapsedTimeInUsecs;
 
     Drivers_open();
     Board_driversOpen();
@@ -95,7 +96,7 @@ void *mcspi_loopback_main(void *args)
     DebugP_log("----------------------------------------------------------\r\n");
     DebugP_log("Data Width \tData Length \tTransfer Time (micro sec)\r\n");
     DebugP_log("%u\t\t%u\t\t%5.2f\r\n", spiTransaction.dataSize, APP_MCSPI_MSGSIZE,
-                        (float)elapsedTimeInUsecs / APP_MCSPI_TRANSFER_LOOPCOUNT);
+                        (Float32)(elapsedTimeInUsecs / (uint32_t)APP_MCSPI_TRANSFER_LOOPCOUNT));
     DebugP_log("----------------------------------------------------------\r\n\n");
 
     if((SystemP_SUCCESS != transferOK) ||
